requirements.txt
asn1crypto==1.5.1
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
exceptiongroup==1.2.2
filelock==3.17.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
paramiko==3.5.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pytest==8.3.4
pytz==2025.1
requests==2.32.3
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
sshtunnel==0.4.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==1.26.20
